Fix issue with casting `primitive.D` to `Template` by correctly unmarshalling
Restructure a bunch of code into more relevant areas of the codebase
