+++ internal/models/answer.go
@@ -16,11 +16,11 @@ type Answer struct {
 	QuestionAnswers []QuestionAnswer   `bson:"questions,omitempty"`
 }
 
-func (a Answer) FromForm(templateID primitive.ObjectID, username string, questionViews []QuestionView) Answer {
+func (a Answer) FromForm(templateID primitive.ObjectID, username string, questionAnswers []QuestionAnswer) Answer {
 	a.TemplateID = templateID
 	a.Username = username
 	a.CreatedAt = time.Now()
-	a.QuestionAnswers = QuestionAnswersFromForm(questionViews)
+	a.QuestionAnswers = questionAnswers
 
 	return a
 }
